requirements.txt
pandas
numpy
matplotlib
scikit-learn
xgboost
pyarrow
pyyaml
shap
